Day 7 Kotlin: some cleanup

+++ src/com/tenebraex8/aoc2021/Extensions.kt
@@ -78,11 +78,11 @@ fun Any.secondSolution(){
     this.doPrintln("Solution for Part 2: ")
 }
 
-fun Iterable<Int>.median(): Double {
+fun Iterable<Int>.median(): Int {
     val sorted = this.sorted()
     if (sorted.size % 2 == 0) {
-        return ((sorted[sorted.size / 2] + sorted[sorted.size / 2 - 1]) / 2).toDouble()
+        return ((sorted[sorted.size / 2] + sorted[sorted.size / 2 - 1]) / 2)
     } else {
-        return (sorted[sorted.size / 2]).toDouble()
+        return (sorted[sorted.size / 2])
     }
 }
